backend/requirements.txt
 langchain
 faiss-cpu
 python-dotenv
 openai
 textblob
 langchain-community
 tiktoken
 sentence-transformers
